fix(frontend): keep image tests independent of generated types
